Tela Login/Register e Estilização

+++ backend/routes/products.py
@@ -1,27 +1,32 @@
 from flask import Blueprint, request, jsonify
-from config import app, mysql
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from config import mysql
 
 produtos_bp = Blueprint('produtos', __name__)
 
-# Endpoint cadastrar um produto
 
-
-@app.route('/produtos', methods=['POST'])
+@produtos_bp.route('/produtos', methods=['POST'])
+@jwt_required()
 def cadastrar_produto():
+    usuario_id = int(get_jwt_identity())
+    if usuario_id != 2:
+        return jsonify({"erro": "Apenas o usuário autorizado pode adicionar produtos."}), 403
+
     data = request.json
     nome = data['nome']
     categoria = data['categoria']
     quantidade = data['quantidade']
     preco = data['preco']
+
     if not nome or not categoria or not quantidade or not preco:
         return jsonify({"erro": "Todos os campos (nome, categoria, quantidade e preço são obrigatórios)"}), 400
 
     cur = mysql.connection.cursor()
     cur.execute("SELECT * FROM produtos WHERE nome = %s", (nome,))
-    produto_existente = cur.fetchone()
-    if produto_existente:
+    if cur.fetchone():
         cur.close()
         return jsonify({"erro": "Um produto com este nome já está cadastrado!"}), 409
+
     cur.execute(
         "INSERT INTO produtos (nome, categoria, quantidade, preco) VALUES (%s, %s, %s, %s)",
         (nome, categoria, quantidade, preco)
@@ -31,11 +36,8 @@ def cadastrar_produto():
 
     return jsonify({"mensagem": "Produto cadastrado com sucesso!"})
 
-# Endpoint listar produtos
 
-
-@app.route('/produtos', methods=['GET'])
-@app.route('/produtos', methods=['GET'])
+@produtos_bp.route('/produtos', methods=['GET'])
 def listar_produtos():
     id = request.args.get('id')
     nome = request.args.get('nome')
@@ -67,8 +69,8 @@ def listar_produtos():
     return jsonify(lista_produtos)
 
 
-@app.route('/produtos/<int:id>', methods=['GET'])
-def busca_produtos(id):
+@produtos_bp.route('/produtos/<int:id>', methods=['GET'])
+def busca_produto(id):
     cur = mysql.connection.cursor()
     cur.execute("SELECT * FROM produtos WHERE id = %s", (id,))
     produto = cur.fetchone()
@@ -79,40 +81,53 @@ def busca_produtos(id):
         return jsonify({"erro": "Produto não encontrado"}), 404
 
 
-# Endpoint editar produtos
+@produtos_bp.route('/produtos/<int:id>', methods=['PUT'])
+@jwt_required()
+def atualizar_produto(id):
+    usuario_id = int(get_jwt_identity())
+    if usuario_id != 2:
+        return jsonify({"erro": "Apenas o usuário autorizado pode editar produtos."}), 403
 
-@app.route('/produtos/<int:id>', methods=['PUT'])
-def atualizar_produtos(id):
     data = request.get_json()
     nome = data.get('nome')
     categoria = data.get('categoria')
     quantidade = data.get('quantidade')
     preco = data.get('preco')
+
     cur = mysql.connection.cursor()
     cur.execute("SELECT * FROM produtos WHERE id = %s", (id,))
     produto = cur.fetchone()
 
     if not produto:
         return jsonify({'erro': 'Produto não encontrado'}), 404
+
     cur.execute("SELECT * FROM produtos WHERE nome = %s AND id != %s", (nome, id))
     if cur.fetchone():
         return jsonify({"erro": 'Já existe um produto com esse nome'}), 400
-    cur.execute("UPDATE produtos SET nome = %s, categoria = %s, quantidade = %s, preco = %s WHERE id = %s", (nome, categoria, quantidade, preco, id))
+
+    cur.execute("UPDATE produtos SET nome = %s, categoria = %s, quantidade = %s, preco = %s WHERE id = %s",
+                (nome, categoria, quantidade, preco, id))
     mysql.connection.commit()
     cur.close()
-    return jsonify({'mensagem': 'Produto Atualizado com sucesso'})
+
+    return jsonify({'mensagem': 'Produto atualizado com sucesso'})
 
 
-# Endpoint excluir produtos
-
-@app.route('/produtos/<int:id>', methods=['DELETE',])
+@produtos_bp.route('/produtos/<int:id>', methods=['DELETE'])
+@jwt_required()
 def deletar_produto(id):
+    usuario_id = int(get_jwt_identity())
+    if usuario_id != 2:
+        return jsonify({"erro": "Apenas o usuário autorizado pode excluir produtos."}), 403
+
     cur = mysql.connection.cursor()
     cur.execute("SELECT * FROM produtos WHERE id = %s", (id,))
     produto = cur.fetchone()
+
     if not produto:
         cur.close()
         return jsonify({"erro": "Produto não encontrado ou já foi removido"}), 404
+
     cur.execute("DELETE FROM produtos WHERE id = %s", (id,))
     mysql.connection.commit()
     cur.close()
@@ -120,10 +135,6 @@ def deletar_produto(id):
     return jsonify({"mensagem": f"Produto com id {id} removido com sucesso!"})
 
 
-@app.route('/produtos', methods=['DELETE'])
+@produtos_bp.route('/produtos', methods=['DELETE'])
 def deletar_sem_id():
     return jsonify({"erro": "É necessário informar o ID do produto que deseja excluir!"}), 400
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
