+++ frontend/src/components/DeleteProductButton.js
@@ -1,13 +1,16 @@
 import { useState } from 'react';
-import instance from '../services/api'
+import instance from '../services/api';
 
 const DeleteProductButton = ({ id, onDelete }) => {
     const [errorMessage, setErrorMessage] = useState("");
 
     const excluirProduto = () => {
+        const confirmar = window.confirm("Tem certeza que deseja excluir esse produto?");
+        if (!confirmar) return;
+
         instance.delete(`/produtos/${id}`)
             .then(() => {
-                onDelete(id); 
+                onDelete(id);
             })
             .catch(error => {
                 setErrorMessage("Erro: Produto não encontrado.");
